+++ src/main/java/net/sf/mpxj/BaselineManager.java
@@ -50,8 +50,14 @@ class BaselineManager
    /**
     * Use the supplied baseline project to set the baselineN cost, duration, finish,
     * fixed cost accrual, fixed cost, start and work attributes for the tasks
-    * in the supplied project. The supplied TaskField instance is the key used
-    * to match tasks in the current project with tasks in the baseline project.
+    * in the supplied project.
+    *
+    * The supplied keyFunction is used to generate the key
+    * used to connect tasks from the current and baseline schedules. This key should
+    * be unique for each task in the schedule. Instances where the key is not unique
+    * will result in an incorrect baseline being applied to a task in the
+    * current schedule.
+    *
     * The index argument selects which of the 10 baselines to populate. Passing
     * an index of 0 populates the main baseline.
     *
@@ -63,7 +69,7 @@ class BaselineManager
    public void populateBaseline(ProjectFile project, ProjectFile baseline, int index, Function<Task, Object> keyFunction)
    {
       TaskField[] baselineFields = getBaselineFields(index);
-      Map<Object, Task> map = baseline.getTasks().stream().filter(t -> keyFunction.apply(t) != null).collect(Collectors.toMap(t -> keyFunction.apply(t), t -> t));
+      Map<Object, Task> map = baseline.getTasks().stream().filter(t -> keyFunction.apply(t) != null).collect(Collectors.toMap(t -> keyFunction.apply(t), t -> t, (u,v) -> null));
       project.getTasks().forEach(t -> populateBaseline(t, map.get(keyFunction.apply(t)), baselineFields));
    }