+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 9.5.2 (git master)
 * Add Must Finish By date to project properties.
 * Rename ProjectFile.getStartDate() and ProjectFile.getFinishDate() methods for clarity. Original method names are marked as deprecated
+* Ensure that all activities in a PMXML file have a CalendarID attribute to ensure compatibility with older versions of P6.
 
 ## 9.5.1 (01/07/2021)
 * When applying a baseline using ProjectFile.setBaseline, gracefully handle duplicate task key values.